# Speedup debezium parse