Before:
```
BenchmarkReceiverTest
BenchmarkReceiverTest/parse
BenchmarkReceiverTest/parse/tz
BenchmarkReceiverTest/parse/tz-10      	   53989	     22082 ns/op	  15.58 MB/s	    9008 B/op	      55 allocs/op
BenchmarkReceiverTest/parse/no-tz
BenchmarkReceiverTest/parse/no-tz-10   	   54669	     21947 ns/op	  15.67 MB/s	    9007 B/op	      55 allocs/op
PASS
```

After
```
BenchmarkReceiverTest
BenchmarkReceiverTest/parse
BenchmarkReceiverTest/parse/citext
BenchmarkReceiverTest/parse/citext-10  	  134011	      8474 ns/op	  39.30 MB/s	   12149 B/op	      36 allocs/op
BenchmarkReceiverTest/parse/tz
BenchmarkReceiverTest/parse/tz-10      	  130683	      9110 ns/op	  37.76 MB/s	   12679 B/op	      36 allocs/op
BenchmarkReceiverTest/parse/no-tz
BenchmarkReceiverTest/parse/no-tz-10   	  135462	      8913 ns/op	  38.60 MB/s	   12679 B/op	      36 allocs/op
```

Main gain from faster json parse, total gain x3